\name{gRcox}
\alias{gRcox}
\title{
  The package `gRcox': summary information
}
\description{
  This package facilitates statistical inference in RCON and RCOR
  models. See 'References' for references
}
\section{Details}{
  \itemize{
    \item The package is part of the gR initiative.
  }
  The package is intended as a contribution to the gR-project 
  derscribed by Lauritzen (2002).
}
\section{Authors}{
  Søren Højsgaard, Statistics and Decision Theory Research Unit, Danish Institute of
  Agricultural Sciences, DK-8830 Tjele, Denmark
}
%\section{Acknowledgements}{
%  Thanks to David Edwards, the creator of the MIM program.
%  
%}  
\references{
  Lauritzen, S. L. (2002). gRaphical Models in R. \emph{R News}, 3(2)39.
} 
\keyword{multivariate}
\keyword{models}
\keyword{graphs}
